requirements.txt
Python>=3.0.0
numpy>=1.9.2
pandas>=0.19
torchvision>=0.13.1
opencv-python>=4.6.0.66
pytorch>=1.12.0
paddleocr>=2.6.1.0
paddlepaddle>=2.4.0
glob2==0.7